+++ appsrc/models/mDebug.js
@@ -1,4 +1,4 @@
-// src/models/mDebug.js v0.0.2-2
+// appsrc/models/mDebug.js v0.0.3-3
 /*
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -10,7 +10,9 @@ var Env = 'DEV'
 var init = function (tt) {
 
     // Add each module to debug
-    ['mPatientList', 'vPatientAdd', 'vPatientEdit'].map(mod => {
+    [ 'mPatientList',
+      'vPatientList', 'vPatientAdd', 'vPatientEdit'
+    ].map(mod => {
         var dir
         if (mod[0] === 'm') { dir = 'models' }
         if (mod[0] === 'v') { dir = 'views' }